[#27] Move survey question display type ui to survey package
